// api/model/fcCarPartType.ts
/**
 * CarPartStore-API
 * API des CarPartStore-Backends
 *
 * The version of the OpenAPI document: 1.0
 * Contact: ata@hbt.de
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export type FcCarPartType = 'REIFEN_FELGEN' | 'SCHEIBEN' | 'SONSTIGES' | 'KAROSSERIE' | 'INNEN_AUSSTATTUNG' | 'SCHRAUBEN' | 'MOTOR' | 'FLUESSIGKEIT' | 'FARBE';

export const FcCarPartType = {
    ReifenFelgen: 'REIFEN_FELGEN' as FcCarPartType,
    Scheiben: 'SCHEIBEN' as FcCarPartType,
    Sonstiges: 'SONSTIGES' as FcCarPartType,
    Karosserie: 'KAROSSERIE' as FcCarPartType,
    InnenAusstattung: 'INNEN_AUSSTATTUNG' as FcCarPartType,
    Schrauben: 'SCHRAUBEN' as FcCarPartType,
    Motor: 'MOTOR' as FcCarPartType,
    Fluessigkeit: 'FLUESSIGKEIT' as FcCarPartType,
    Farbe: 'FARBE' as FcCarPartType
};